DEV: edit FXrates
