+++ src/lib.rs
@@ -31,7 +31,8 @@ use calendars::calendar::{Cal, Modifier, RollDay, UnionCal};
 use calendars::calendar_py::get_calendar_by_name_py;
 
 pub mod fx;
-use fx::rates::{Ccy, FXRate, FXRates};
+use fx::rates::ccy::Ccy;
+use fx::rates::{FXRate, FXRates};
 
 #[pymodule]
 fn rs(m: &Bound<'_, PyModule>) -> PyResult<()> {
